requirements.txt
black==22.12.0
Django==4.1.4
django-flexible-subscriptions==0.15.1
flake8==6.0.0
isort==5.10.1
pytest==7.2.0
pytest-django==4.5.2
tox==4.0.8
venmo-api==0.3.1
